+++ app.json
@@ -1,10 +1,10 @@
 {
   "expo": {
-    "name": "xpressfoods",
-    "slug": "xpressfoods",
+    "name": "Jirani",
+    "slug": "Jirani",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/images/icon.png",
+    "icon": "./assets/images/app-icon.png",
     "scheme": "myapp",
     "userInterfaceStyle": "automatic",
     "splash": {
