+++ package.json
@@ -43,6 +43,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.74.3",
     "react-native-autocomplete-dropdown": "^4.0.0-rc.5",
+    "react-native-element-dropdown": "^2.12.1",
     "react-native-elements": "^3.4.3",
     "react-native-feather": "^1.1.2",
     "react-native-gesture-handler": "~2.16.1",
@@ -50,6 +51,8 @@
     "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
+    "react-native-searchable-dropdown": "^1.1.3",
+    "react-native-searchable-dropdown-kj": "^1.9.3",
     "react-native-svg": "^15.5.0",
     "react-native-svg-transformer": "^1.5.0",
     "react-native-tab-view": "^3.5.2",
