+++ src/app/(admin)/carwashAdmin/paymentHistory.tsx
@@ -124,6 +124,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: "#f0f0f0",
+    marginTop: 40,
   },
   header: {
     alignItems: "center",
