+++ src/app/(admin)/index.tsx
@@ -1,5 +1,5 @@
 import { Redirect } from "expo-router";
 
 export default function TabIndex() {
-  return <Redirect href={"/(admin)/menu/"} />;
+  return <Redirect href={"/(admin)/carwashAdmin/"} />;
 }
