+++ src/app/index.tsx
@@ -22,14 +22,15 @@ const index = () => {
   /* if (!isAdmin) {
     return <Redirect href={"/(user)/carwash"} />;
   } */
+  console.log(isAdmin);
 
   if (isAdmin) {
     return (
       <View style={{ flex: 1, justifyContent: "center", padding: 10 }}>
-        <Link href={"/(user)"} asChild>
+        <Link href={`/(user)/${profile.admin_app}`} asChild>
           <Button text="User" />
         </Link>
-        <Link href={"/(admin)"} asChild>
+        <Link href={`/(admin)/${profile.admin_app}Admin`} asChild>
           <Button text="Admin" />
         </Link>
       </View>
