+++ src/components/admin/_orders/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack, Link } from "expo-router";
+
+export default function MenuStack() {
+  return (
+    <Stack>
+      <Stack.Screen name="list" options={{ headerShown: false }} />
+    </Stack>
+  );
+}
