+++ src/components/admin/_orders/list/archive.tsx
@@ -1,0 +1,26 @@
+import { Text, FlatList, ActivityIndicator } from "react-native";
+import OrderListItem from "@/src/components/OrderListItem";
+import { useAdminOrderList } from "@/src/api/orders";
+
+export default function OrdersScreen() {
+  const {
+    data: orders,
+    isLoading,
+    error,
+  } = useAdminOrderList({ archived: true });
+
+  if (isLoading) {
+    return <ActivityIndicator />;
+  }
+  if (error) {
+    return <Text>failed to fetch products</Text>;
+  }
+
+  return (
+    <FlatList
+      data={orders}
+      renderItem={({ item }) => <OrderListItem order={item} />}
+      contentContainerStyle={{ padding: 10, gap: 10 }}
+    />
+  );
+}
