+++ src/components/admin/menu/[id].tsx
@@ -1,0 +1,104 @@
+import products from "@/assets/data/products";
+import { useLocalSearchParams, Stack, useRouter, Link } from "expo-router";
+import {
+  View,
+  Text,
+  Image,
+  StyleSheet,
+  Pressable,
+  ActivityIndicator,
+} from "react-native";
+import { useState } from "react";
+import Button from "@/src/components/Button";
+import { useCart } from "@/src/providers/CartProvider";
+import { PizzaSize } from "@/src/types";
+import { FontAwesome } from "@expo/vector-icons";
+import Colors from "@/src/constants/Colors";
+import { useProduct } from "@/src/api/products";
+import RemoteImage from "@/src/components/RemoteImage";
+
+const sizes: PizzaSize[] = ["S", "M", "L", "XL"];
+
+const ProductDetailsScreen = () => {
+  const { id: idString } = useLocalSearchParams();
+
+  const id = parseFloat(typeof idString === "string" ? idString : "");
+
+  const { data: product, error, isLoading } = useProduct(id);
+
+  const { addItem } = useCart();
+
+  const router = useRouter();
+
+  const [selectedSize, setSelectedSize] = useState<PizzaSize>("M");
+
+  if (!product) {
+    return <Text>Product Not Found</Text>;
+  }
+
+  if (isLoading) {
+    return <ActivityIndicator />;
+  }
+  if (error) {
+    return <Text>failed to fetch products</Text>;
+  }
+
+  const addToCart = () => {
+    addItem(product, selectedSize);
+    router.push("/cart");
+  };
+
+  return (
+    <View style={styles.container}>
+      <Stack.Screen
+        options={{
+          title: "Menu",
+          headerRight: () => (
+            <Link href={`/(admin)/menu/create?id=${id}`} asChild>
+              <Pressable>
+                {({ pressed }) => (
+                  <FontAwesome
+                    name="pencil"
+                    size={25}
+                    color={Colors.light.tint}
+                    style={{ marginRight: 15, opacity: pressed ? 0.5 : 1 }}
+                  />
+                )}
+              </Pressable>
+            </Link>
+          ),
+        }}
+      />
+      <Stack.Screen options={{ title: product?.name }} />
+      <RemoteImage
+        path={product.image}
+        fallback={product.image}
+        style={styles.image}
+      />
+
+      <Text style={styles.title}>{product.name}</Text>
+      <Text style={styles.price}>${product.price}</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "white",
+    flex: 1,
+    padding: 10,
+  },
+  image: {
+    width: "100%",
+    aspectRatio: 1,
+  },
+  title: {
+    fontSize: 20,
+  },
+  price: {
+    fontWeight: "bold",
+    fontSize: 18,
+  },
+});
+
+export default ProductDetailsScreen;
