--- src/components/orders/_layout.tsx
@@ -1,9 +1,0 @@
-import { Stack, Link } from "expo-router";
-
-export default function MenuStack() {
-  return (
-    <Stack>
-      <Stack.Screen name="index" options={{ title: "Orders" }} />
-    </Stack>
-  );
-}
