+++ src/components/user/menu/[id] copy.tsx
@@ -1,0 +1,110 @@
+import { useLocalSearchParams, Stack, useRouter } from "expo-router";
+import { View, Text, Image, StyleSheet, Pressable } from "react-native";
+import { useState } from "react";
+import Button from "@/src/components/Button";
+import { useCart } from "@/src/providers/CartProvider";
+import { PizzaSize } from "@/src/types";
+import { useProduct } from "@/src/api/products";
+import { ActivityIndicator } from "react-native";
+import RemoteImage from "@/src/components/RemoteImage";
+
+const sizes: PizzaSize[] = ["S", "M", "L", "XL"];
+
+const ProductDetailsScreen = () => {
+  const { id: idString } = useLocalSearchParams();
+
+  const id = parseFloat(typeof idString === "string" ? idString : "");
+
+  const { data: product, error, isLoading } = useProduct(id);
+
+  const { addItem } = useCart();
+
+  const router = useRouter();
+
+  const [selectedSize, setSelectedSize] = useState<PizzaSize>("M");
+
+  if (isLoading) {
+    return <ActivityIndicator />;
+  }
+  if (error) {
+    return <Text>failed to fetch products</Text>;
+  }
+
+  const addToCart = () => {
+    addItem(product, selectedSize);
+    router.push("/cart");
+  };
+
+  return (
+    <View style={styles.container}>
+      <Stack.Screen options={{ title: product?.name }} />
+      <RemoteImage path={product.image} fallback="" style={styles.image} />
+
+      <Text>Select size</Text>
+      <View style={styles.sizes}>
+        {sizes.map((size) => (
+          <Pressable
+            onPress={() => {
+              setSelectedSize(size);
+            }}
+            style={[
+              styles.sizes,
+              {
+                backgroundColor: selectedSize === size ? "gainsboro" : "white",
+              },
+            ]}
+            key={size}
+          >
+            <Text
+              style={[
+                styles.sizeText,
+                { color: selectedSize === size ? "black" : "grey" },
+              ]}
+            >
+              {size}
+            </Text>
+          </Pressable>
+        ))}
+      </View>
+
+      <Text style={styles.price}>${product.price}</Text>
+      <Button onPress={addToCart} text="Add to Cart" />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "white",
+    flex: 1,
+    padding: 10,
+  },
+  image: {
+    width: "100%",
+    aspectRatio: 1,
+  },
+  price: {
+    fontWeight: "bold",
+    fontSize: 18,
+    marginTop: "auto",
+  },
+  sizes: {
+    flexDirection: "row",
+    justifyContent: "space-around",
+    marginVertical: 10,
+  },
+  size: {
+    backgroundColor: "gainsboro",
+    width: 50,
+    aspectRatio: 1,
+    borderRadius: 25,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  sizeText: {
+    fontSize: 20,
+    fontWeight: "500",
+  },
+});
+
+export default ProductDetailsScreen;
