+++ src/components/user/orders/index.tsx
@@ -1,0 +1,23 @@
+import { Text, FlatList, ActivityIndicator } from "react-native";
+
+import OrderListItem from "@/src/components/OrderListItem";
+import { useMyOrderList } from "@/src/api/orders";
+
+export default function OrdersScreen() {
+  const { data: orders, isLoading, error } = useMyOrderList();
+
+  if (isLoading) {
+    return <ActivityIndicator />;
+  }
+  if (error) {
+    return <Text>failed to fetch products</Text>;
+  }
+
+  return (
+    <FlatList
+      data={orders}
+      renderItem={({ item }) => <OrderListItem order={item} />}
+      contentContainerStyle={{ padding: 10, gap: 10 }}
+    />
+  );
+}
